+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 - Prevent cancellation if process is at 99%, ensuring the process is not interrupted at the last moment.
 - Improved process handling for subpprocess calls, ensuring better management of subprocesses.
 - Improved PDF handling, ignoring empty pages/chapters and better chapter handling.
-- Added `.opus` as output format for generated audio files.
+- Added `Save in a project folder with metadata` option in the book handler, allowing users to save the converted items in a project folder with available metadata files. Useful if you want to work with the converted files in the future, issue mentioned by @Darthagnon in #15
+- Added `Go to folder` button in input box, allowing users to open the folder containing the converted file.
+- Added `.opus` as output format for generated audio files, which is a more efficient format for audio files.
 - Added "Playing..." indicator for "Preview" button in the voice mixer.
 
 # v1.0.6